add interception route
