+++ pruebas/01-reading-list/blasferna/src/app/@modal/(.)books/[isbn]/default.jsx
@@ -1,0 +1,3 @@
+export default function Default(){
+    return null;
+}
