+++ pruebas/01-reading-list/blasferna/src/app/@modal/(.)books/[isbn]/page.jsx
@@ -1,0 +1,11 @@
+import Modal from "@/components/modal";
+import BookDetail from "@/components/book-detail";
+
+export default function BookDetailModalPage({ params }) {
+  const { isbn } = params;
+  return (
+    <Modal>
+      <BookDetail isbn={isbn}></BookDetail>
+    </Modal>
+  );
+}
